kaggle_pipeline/service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
insightface==0.7.3
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.3.0
cloudpickle==3.0.0
